Initial 2FA commit

+++ backend/middleware/requireAuth.js
@@ -1,18 +1,31 @@
 import AuthToken from "../models/AuthToken.js";
 import User from "../models/User.js";
+import crypto from "crypto";
 
 const requireAuth = async (req, res, next) => {
-    const token = req.cookies.token
-    if (!token){
-        return res.status(401).json({errorMessage: "This is a protected route. Please authenticate first" , responseImage: "https://www.meme-arsenal.com/memes/f7f7f1287442780dfd544539f6dc2f15.jpg"})
+    const tokenCookie = req.cookies.token;
+    if (!tokenCookie){
+        return res.status(401).json({authenticated: false, errorMessage: "No authentication cookie recieved. Authentication failed."})
     }
-    const authRecord = await AuthToken.findOne({token})
-    if (!authRecord){
-        return res.status(401).json({errorMessage: "Invalid token", responseImage: "https://www.meme-arsenal.com/memes/f7f7f1287442780dfd544539f6dc2f15.jpg"})
+    const [token, selector] = tokenCookie.split(":")
+    
+    if (!token || !selector) {
+        return res.status(401).json({authenticated: false, errorMessage: "Invalid authentication cookie format. Authentication failed." });
     }
+    const authRecord = await AuthToken.findOne({selector})
+
+    if (!authRecord) {
+        return res.status(401).json({authenticated: false, errorMessage: "Authentication cookie doesn't match records. Authentication failed" });
+    }
+
     if (authRecord.expiresAt < new Date()){
         await AuthToken.deleteOne({_id: authRecord._id})
-        return res.status(401).json({errorMessage: "Expired token", responseImage: "https://www.meme-arsenal.com/memes/f7f7f1287442780dfd544539f6dc2f15.jpg"})
+        return res.status(401).json({authenticated: false, errorMessage: "Authentication token expired. Please login again."})
+    }
+
+    const hashed = crypto.createHash("sha256").update(token).digest("hex");
+    if (hashed != authRecord.hashedToken){
+        return res.status(401).json({authenticated: false, errorMessage: "Invalid token" });
     }
     const userId = authRecord.userId;
     const user = await User.findOne({_id: userId})
